requirements.txt
discord.py>=2.3.0
ccxt>=4.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
web3[c-kzg]>=6.0.0 
matplotlib>=3.8.0
numpy<2.0
tenacity>=8.0.0
py-solc-x>=2.0.0
